# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SpaceLink API Documents",
        "description": "Hệ thống API backend viết bằng Laravel 12",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Danh sách địa chỉ của user hiện tại",
                "operationId": "dc6b190d3f07db7ed032fa181dc45150",
                "responses": {
                    "200": {
                        "description": "Danh sách địa chỉ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "fullname": {
                                                        "type": "string"
                                                    },
                                                    "phone": {
                                                        "type": "string"
                                                    },
                                                    "province": {
                                                        "type": "string"
                                                    },
                                                    "district": {
                                                        "type": "string"
                                                    },
                                                    "ward": {
                                                        "type": "string"
                                                    },
                                                    "address_detail": {
                                                        "type": "string"
                                                    },
                                                    "is_default": {
                                                        "type": "boolean"
                                                    },
                                                    "address_type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "home",
                                                            "office",
                                                            "other"
                                                        ]
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Address"
                ],
                "summary": "Thêm địa chỉ mới",
                "operationId": "6344fc023028e43a5490f1c6372bd391",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullname",
                                    "phone",
                                    "province",
                                    "district",
                                    "ward",
                                    "address_detail"
                                ],
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0912345678"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Hồ Chí Minh"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Quận 1"
                                    },
                                    "ward": {
                                        "type": "string",
                                        "example": "Phường Bến Nghé"
                                    },
                                    "address_detail": {
                                        "type": "string",
                                        "example": "123 Nguyễn Huệ"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "address_type": {
                                        "type": "string",
                                        "enum": [
                                            "home",
                                            "office",
                                            "other"
                                        ],
                                        "example": "home"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thêm thành công"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/addresses/{id}": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Chi tiết 1 địa chỉ",
                "operationId": "2a46bc21bf18060f0ec3a4fff2debe57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Address"
                ],
                "summary": "Cập nhật địa chỉ",
                "operationId": "b4967e4140f305290dbde0982a3fc7d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullname",
                                    "phone",
                                    "province",
                                    "district",
                                    "ward",
                                    "address_detail"
                                ],
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Nguyen Van B"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0987654321"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Hà Nội"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Quận Hoàn Kiếm"
                                    },
                                    "ward": {
                                        "type": "string",
                                        "example": "Phường Hàng Bạc"
                                    },
                                    "address_detail": {
                                        "type": "string",
                                        "example": "45 Hàng Ngang"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "address_type": {
                                        "type": "string",
                                        "enum": [
                                            "home",
                                            "office",
                                            "other"
                                        ],
                                        "example": "office"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Address"
                ],
                "summary": "Xóa địa chỉ",
                "operationId": "eedadf79fef97e5ee1af0b11c8856d8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Nhập Token theo định dạng: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}